+++ lib/src/components/switch.dart
@@ -149,9 +149,9 @@ class ShadSwitch extends StatefulWidget {
   /// {@template ShadSwitch.padding}
   /// Padding between the switch itself and its label.
   ///
-  /// Defaults to `EdgeInsets.only(left: 8)`.
+  /// Defaults to `EdgeInsetsDirectional.only(start: 8)`.
   /// {@endtemplate}
-  final EdgeInsets? padding;
+  final EdgeInsetsGeometry? padding;
 
   @override
   State<ShadSwitch> createState() => _ShadSwitchState();
@@ -177,6 +177,14 @@ class _ShadSwitchState extends State<ShadSwitch>
     if (oldWidget.value != widget.value) {
       controller.animateTo(widget.value ? 1 : 0);
     }
+    if (oldWidget.focusNode != widget.focusNode) {
+      if (widget.focusNode == null) {
+        _focusNode = FocusNode();
+      } else {
+        _focusNode?.dispose();
+        _focusNode = null;
+      }
+    }
   }
 
   @override
@@ -197,6 +205,8 @@ class _ShadSwitchState extends State<ShadSwitch>
   Widget build(BuildContext context) {
     assert(debugCheckHasShadTheme(context));
     final theme = ShadTheme.of(context);
+
+    final direction = widget.direction ?? Directionality.of(context);
     final effectiveThumbColor = widget.thumbColor ??
         theme.switchTheme.thumbColor ??
         theme.colorScheme.background;
@@ -212,7 +222,8 @@ class _ShadSwitchState extends State<ShadSwitch>
     final effectiveThumbSize = effectiveHeight - effectiveMargin * 2;
     final transitionStep =
         effectiveWidth - effectiveMargin * 2 - effectiveThumbSize;
-    final effectiveDuration = widget.duration ?? 100.milliseconds;
+    final effectiveDuration =
+        widget.duration ?? theme.switchTheme.duration ?? 100.milliseconds;
 
     final effectiveDecoration =
         (theme.switchTheme.decoration ?? const ShadDecoration())
@@ -225,7 +236,27 @@ class _ShadSwitchState extends State<ShadSwitch>
 
     final effectivePadding = widget.padding ??
         theme.switchTheme.padding ??
-        const EdgeInsets.only(left: 8);
+        const EdgeInsetsDirectional.only(start: 8);
+
+    final effectiveEffects = [
+      switch (direction) {
+        TextDirection.ltr => MoveEffect(
+            begin: Offset.zero,
+            end: Offset(transitionStep, 0),
+            duration: effectiveDuration,
+          ),
+        TextDirection.rtl => MoveEffect(
+            begin: Offset.zero,
+            end: Offset(-transitionStep, 0),
+            duration: effectiveDuration,
+          )
+      },
+    ];
+
+    final keyboardTriggers = <ShortcutActivator>[
+      const SingleActivator(LogicalKeyboardKey.enter),
+      const SingleActivator(LogicalKeyboardKey.space),
+    ];
 
     final switchWidget = Semantics(
       toggled: widget.value,
@@ -234,10 +265,11 @@ class _ShadSwitchState extends State<ShadSwitch>
         disabled: !widget.enabled,
         child: CallbackShortcuts(
           bindings: {
-            const SingleActivator(LogicalKeyboardKey.enter): () {
-              if (!widget.enabled) return;
-              onTap();
-            },
+            for (final trigger in keyboardTriggers)
+              trigger: () {
+                if (!widget.enabled) return;
+                onTap();
+              },
           },
           child: ShadFocusable(
             focusNode: focusNode,
@@ -260,23 +292,22 @@ class _ShadSwitchState extends State<ShadSwitch>
                     child: Animate(
                       controller: controller,
                       autoPlay: false,
-                      effects: [
-                        MoveEffect(
-                          begin: Offset.zero,
-                          end: Offset(transitionStep, 0),
-                          duration: effectiveDuration,
-                        ),
-                      ],
-                      child: SizedBox(
-                        width: effectiveThumbSize,
-                        height: effectiveThumbSize,
-                        child: ShadDecorator(
-                          decoration: ShadDecoration(
-                            color: effectiveThumbColor,
-                            shape: BoxShape.circle,
-                            merge: false,
+                      effects: effectiveEffects,
+                      child: Row(
+                        textDirection: direction,
+                        children: [
+                          SizedBox(
+                            width: effectiveThumbSize,
+                            height: effectiveThumbSize,
+                            child: ShadDecorator(
+                              decoration: ShadDecoration(
+                                color: effectiveThumbColor,
+                                shape: BoxShape.circle,
+                                merge: false,
+                              ),
+                            ),
                           ),
-                        ),
+                        ],
                       ),
                     ),
                   ),
@@ -295,7 +326,7 @@ class _ShadSwitchState extends State<ShadSwitch>
       child: GestureDetector(
         onTap: widget.onChanged == null ? null : onTap,
         child: Row(
-          textDirection: widget.direction,
+          textDirection: direction,
           mainAxisSize: MainAxisSize.min,
           children: [
             switchWidget,