+++ lib/src/theme/components/dialog.dart
@@ -37,6 +37,7 @@ class ShadDialogTheme {
     this.scrollable,
     this.scrollPadding,
     this.actionsGap,
+    this.useSafeArea,
   });
 
   final bool merge;
@@ -57,7 +58,7 @@ class ShadDialogTheme {
   final bool? expandActionsWhenTiny;
 
   /// {@macro ShadDialog.padding}
-  final EdgeInsets? padding;
+  final EdgeInsetsGeometry? padding;
 
   /// {@macro ShadDialog.gap}
   final double? gap;
@@ -117,10 +118,13 @@ class ShadDialogTheme {
   final bool? scrollable;
 
   /// {@macro ShadDialog.scrollPadding}
-  final EdgeInsets? scrollPadding;
+  final EdgeInsetsGeometry? scrollPadding;
 
   /// {@macro ShadDialog.actionsGap}
   final double? actionsGap;
+
+  /// {@macro ShadDialog.useSafeArea}
+  final bool? useSafeArea;
 
   static ShadDialogTheme lerp(
     ShadDialogTheme a,
@@ -136,7 +140,7 @@ class ShadDialogTheme {
       radius: BorderRadius.lerp(a.radius, b.radius, t),
       backgroundColor: Color.lerp(a.backgroundColor, b.backgroundColor, t),
       expandActionsWhenTiny: b.expandActionsWhenTiny,
-      padding: EdgeInsets.lerp(a.padding, b.padding, t),
+      padding: EdgeInsetsGeometry.lerp(a.padding, b.padding, t),
       gap: lerpDouble(a.gap, b.gap, t),
       animateIn: b.animateIn,
       animateOut: b.animateOut,
@@ -157,8 +161,10 @@ class ShadDialogTheme {
       mainAxisAlignment: t < 0.5 ? a.mainAxisAlignment : b.mainAxisAlignment,
       crossAxisAlignment: t < 0.5 ? a.crossAxisAlignment : b.crossAxisAlignment,
       scrollable: t < 0.5 ? a.scrollable : b.scrollable,
-      scrollPadding: EdgeInsets.lerp(a.scrollPadding, b.scrollPadding, t),
+      scrollPadding:
+          EdgeInsetsGeometry.lerp(a.scrollPadding, b.scrollPadding, t),
       actionsGap: lerpDouble(a.actionsGap, b.actionsGap, t),
+      useSafeArea: t < 0.5 ? a.useSafeArea : b.useSafeArea,
     );
   }
 
@@ -169,18 +175,18 @@ class ShadDialogTheme {
     BorderRadius? radius,
     Color? backgroundColor,
     bool? expandActionsWhenTiny,
-    EdgeInsets? padding,
+    EdgeInsetsGeometry? padding,
     double? gap,
     List<Effect<dynamic>>? animateIn,
     List<Effect<dynamic>>? animateOut,
     BoxConstraints? constraints,
-    BoxBorder? border,
-    List<BoxShadow>? shadows,
-    bool? removeBorderRadiusWhenTiny,
     Axis? actionsAxis,
     MainAxisSize? actionsMainAxisSize,
     MainAxisAlignment? actionsMainAxisAlignment,
     VerticalDirection? actionsVerticalDirection,
+    BoxBorder? border,
+    List<BoxShadow>? shadows,
+    bool? removeBorderRadiusWhenTiny,
     TextStyle? titleStyle,
     TextStyle? descriptionStyle,
     TextAlign? titleTextAlign,
@@ -189,38 +195,44 @@ class ShadDialogTheme {
     MainAxisAlignment? mainAxisAlignment,
     CrossAxisAlignment? crossAxisAlignment,
     bool? scrollable,
-    EdgeInsets? scrollPadding,
+    EdgeInsetsGeometry? scrollPadding,
     double? actionsGap,
+    bool? useSafeArea,
   }) {
     return ShadDialogTheme(
       merge: merge ?? this.merge,
-      closeIconData: closeIconData,
-      closeIconPosition: closeIconPosition,
-      radius: radius,
-      backgroundColor: backgroundColor,
-      expandActionsWhenTiny: expandActionsWhenTiny,
-      padding: padding,
-      gap: gap,
-      animateIn: animateIn,
-      animateOut: animateOut,
-      constraints: constraints,
-      border: border,
-      shadows: shadows,
-      removeBorderRadiusWhenTiny: removeBorderRadiusWhenTiny,
-      actionsAxis: actionsAxis,
-      actionsMainAxisSize: actionsMainAxisSize,
-      actionsMainAxisAlignment: actionsMainAxisAlignment,
-      actionsVerticalDirection: actionsVerticalDirection,
-      titleStyle: titleStyle,
-      descriptionStyle: descriptionStyle,
-      titleTextAlign: titleTextAlign,
-      descriptionTextAlign: descriptionTextAlign,
-      alignment: alignment,
-      mainAxisAlignment: mainAxisAlignment,
-      crossAxisAlignment: crossAxisAlignment,
-      scrollable: scrollable,
-      scrollPadding: scrollPadding,
-      actionsGap: actionsGap,
+      closeIconData: closeIconData ?? this.closeIconData,
+      closeIconPosition: closeIconPosition ?? this.closeIconPosition,
+      radius: radius ?? this.radius,
+      backgroundColor: backgroundColor ?? this.backgroundColor,
+      expandActionsWhenTiny:
+          expandActionsWhenTiny ?? this.expandActionsWhenTiny,
+      padding: padding ?? this.padding,
+      gap: gap ?? this.gap,
+      animateIn: animateIn ?? this.animateIn,
+      animateOut: animateOut ?? this.animateOut,
+      constraints: constraints ?? this.constraints,
+      actionsAxis: actionsAxis ?? this.actionsAxis,
+      actionsMainAxisSize: actionsMainAxisSize ?? this.actionsMainAxisSize,
+      actionsMainAxisAlignment:
+          actionsMainAxisAlignment ?? this.actionsMainAxisAlignment,
+      actionsVerticalDirection:
+          actionsVerticalDirection ?? this.actionsVerticalDirection,
+      border: border ?? this.border,
+      shadows: shadows ?? this.shadows,
+      removeBorderRadiusWhenTiny:
+          removeBorderRadiusWhenTiny ?? this.removeBorderRadiusWhenTiny,
+      titleStyle: titleStyle ?? this.titleStyle,
+      descriptionStyle: descriptionStyle ?? this.descriptionStyle,
+      titleTextAlign: titleTextAlign ?? this.titleTextAlign,
+      descriptionTextAlign: descriptionTextAlign ?? this.descriptionTextAlign,
+      alignment: alignment ?? this.alignment,
+      mainAxisAlignment: mainAxisAlignment ?? this.mainAxisAlignment,
+      crossAxisAlignment: crossAxisAlignment ?? this.crossAxisAlignment,
+      scrollable: scrollable ?? this.scrollable,
+      scrollPadding: scrollPadding ?? this.scrollPadding,
+      actionsGap: actionsGap ?? this.actionsGap,
+      useSafeArea: useSafeArea ?? this.useSafeArea,
     );
   }
 
@@ -255,6 +267,7 @@ class ShadDialogTheme {
       scrollable: other.scrollable,
       scrollPadding: other.scrollPadding,
       actionsGap: other.actionsGap,
+      useSafeArea: other.useSafeArea,
     );
   }
 
@@ -290,7 +303,8 @@ class ShadDialogTheme {
         other.crossAxisAlignment == crossAxisAlignment &&
         other.scrollable == scrollable &&
         other.scrollPadding == scrollPadding &&
-        other.actionsGap == actionsGap;
+        other.actionsGap == actionsGap &&
+        other.useSafeArea == useSafeArea;
   }
 
   @override
@@ -322,6 +336,7 @@ class ShadDialogTheme {
         crossAxisAlignment.hashCode ^
         scrollable.hashCode ^
         scrollPadding.hashCode ^
-        actionsGap.hashCode;
+        actionsGap.hashCode ^
+        useSafeArea.hashCode;
   }
 }