+++ lib/src/theme/themes/default_theme_variant.dart
@@ -239,6 +239,21 @@ class ShadDefaultThemeVariant extends ShadThemeVariant {
       ),
       duration: Animate.defaultDuration,
       reverseDuration: Duration.zero,
+      hoverStrategies: const ShadHoverStrategies(
+        hover: {
+          ShadHoverStrategy.onTap,
+          ShadHoverStrategy.onLongPressDown,
+          ShadHoverStrategy.onLongPressStart,
+        },
+        unhover: {
+          ShadHoverStrategy.onTap,
+          ShadHoverStrategy.onTapOutside,
+          ShadHoverStrategy.onLongPressUp,
+          ShadHoverStrategy.onLongPressEnd,
+          ShadHoverStrategy.onLongPressCancel,
+        },
+        longPressDuration: kLongPressTimeout,
+      ),
     );
   }
 
@@ -386,7 +401,7 @@ class ShadDefaultThemeVariant extends ShadThemeVariant {
       thumbColor: colorScheme.background,
       uncheckedTrackColor: colorScheme.input,
       checkedTrackColor: colorScheme.primary,
-      padding: const EdgeInsets.only(left: 8),
+      padding: const EdgeInsetsDirectional.only(start: 8),
       decoration: ShadDecoration(
         border: ShadBorder.all(radius: radius.add(radius / 2), width: 0),
         secondaryFocusedBorder:
@@ -401,7 +416,7 @@ class ShadDefaultThemeVariant extends ShadThemeVariant {
       size: 16,
       duration: 100.milliseconds,
       color: colorScheme.primary,
-      padding: const EdgeInsets.only(left: 8),
+      padding: const EdgeInsetsDirectional.only(start: 8),
       checkboxPadding: const EdgeInsets.only(top: 1),
       decoration: ShadDecoration(
         border: ShadBorder.all(
@@ -418,7 +433,6 @@ class ShadDefaultThemeVariant extends ShadThemeVariant {
     return ShadInputTheme(
       style: effectiveTextTheme.muted.copyWith(color: colorScheme.foreground),
       placeholderStyle: effectiveTextTheme.muted,
-      placeholderAlignment: Alignment.topLeft,
       inputPadding: EdgeInsets.zero,
       decoration: ShadDecoration(
         border: ShadBorder.all(
@@ -441,7 +455,7 @@ class ShadDefaultThemeVariant extends ShadThemeVariant {
       circleSize: circleSize,
       duration: 100.milliseconds,
       color: colorScheme.primary,
-      padding: const EdgeInsets.only(left: 8),
+      padding: const EdgeInsetsDirectional.only(start: 8),
       decoration: ShadDecoration(
         shape: BoxShape.circle,
         border: ShadBorder.all(
@@ -474,14 +488,14 @@ class ShadDefaultThemeVariant extends ShadThemeVariant {
       descriptionStyle: effectiveTextTheme.muted.copyWith(
         color: colorScheme.foreground.withValues(alpha: .9),
       ),
-      actionPadding: const EdgeInsets.only(left: 16),
+      actionPadding: const EdgeInsetsDirectional.only(start: 16),
       border: Border.all(color: colorScheme.border),
       shadows: ShadShadows.lg,
       backgroundColor: colorScheme.background,
       crossAxisAlignment: CrossAxisAlignment.center,
       closeIconPosition: const ShadPosition(top: 8, right: 8),
       showCloseIconOnlyWhenHovered: true,
-      padding: const EdgeInsets.fromLTRB(24, 24, 32, 24),
+      padding: const EdgeInsetsGeometry.fromSTEB(24, 24, 32, 24),
     );
   }
 
@@ -497,21 +511,21 @@ class ShadDefaultThemeVariant extends ShadThemeVariant {
       descriptionStyle: effectiveTextTheme.muted.copyWith(
         color: colorScheme.destructiveForeground.withValues(alpha: .9),
       ),
-      actionPadding: const EdgeInsets.only(left: 16),
+      actionPadding: const EdgeInsetsDirectional.only(start: 16),
       border: Border.all(color: colorScheme.border),
       shadows: ShadShadows.lg,
       backgroundColor: colorScheme.destructive,
       crossAxisAlignment: CrossAxisAlignment.center,
       closeIconPosition: const ShadPosition(top: 8, right: 8),
       showCloseIconOnlyWhenHovered: true,
-      padding: const EdgeInsets.fromLTRB(24, 24, 32, 24),
+      padding: const EdgeInsetsDirectional.fromSTEB(24, 24, 32, 24),
     );
   }
 
   @override
   ShadAlertTheme primaryAlertTheme() {
     return ShadAlertTheme(
-      iconPadding: const EdgeInsets.only(right: 12),
+      iconPadding: const EdgeInsetsDirectional.only(end: 12),
       decoration: ShadDecoration(
         border: ShadBorder.all(
           color: colorScheme.border,
@@ -535,7 +549,7 @@ class ShadDefaultThemeVariant extends ShadThemeVariant {
   @override
   ShadAlertTheme destructiveAlertTheme() {
     return ShadAlertTheme(
-      iconPadding: const EdgeInsets.only(right: 12),
+      iconPadding: const EdgeInsetsDirectional.only(end: 12),
       decoration: ShadDecoration(
         border: ShadBorder.all(
           color: colorScheme.destructive,
@@ -729,11 +743,13 @@ class ShadDefaultThemeVariant extends ShadThemeVariant {
       },
       unhover: {
         ShadHoverStrategy.onTapUp,
+        ShadHoverStrategy.onTapOutside,
         ShadHoverStrategy.onTapCancel,
         ShadHoverStrategy.onLongPressUp,
         ShadHoverStrategy.onLongPressEnd,
         ShadHoverStrategy.onLongPressCancel,
       },
+      longPressDuration: kLongPressTimeout,
     );
   }
 
@@ -779,8 +795,8 @@ class ShadDefaultThemeVariant extends ShadThemeVariant {
         constraints: const BoxConstraints(minWidth: 128),
         padding: const EdgeInsets.symmetric(vertical: 4),
         itemPadding: const EdgeInsets.symmetric(horizontal: 4),
-        leadingPadding: const EdgeInsets.only(right: 8),
-        trailingPadding: const EdgeInsets.only(left: 8),
+        leadingPadding: const EdgeInsetsDirectional.only(end: 8),
+        trailingPadding: const EdgeInsetsDirectional.only(start: 8),
         showDelay: const Duration(milliseconds: 100),
         height: 32,
         buttonVariant: ShadButtonVariant.ghost,
@@ -994,7 +1010,6 @@ class ShadDefaultThemeVariant extends ShadThemeVariant {
     return ShadTextareaTheme(
       style: effectiveTextTheme.muted.copyWith(color: colorScheme.foreground),
       placeholderStyle: effectiveTextTheme.muted,
-      placeholderAlignment: Alignment.topLeft,
       inputPadding: EdgeInsets.zero,
       decoration: ShadDecoration(
         border: ShadBorder.all(